+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/cases/DocumentController.java
@@ -1,0 +1,53 @@
+package uk.gov.hmcts.reform.civil.controllers.cases;
+
+import io.swagger.annotations.Api;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+import uk.gov.hmcts.reform.civil.controllers.testingsupport.CamundaRestEngineClient;
+import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
+import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
+import uk.gov.hmcts.reform.civil.service.docmosis.sealedclaim.SealedClaimFormGeneratorForSpec;
+
+import javax.validation.constraints.NotNull;
+
+
+@Api
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping(
+    path = "/case/document"
+)
+public class DocumentController {
+
+    private final CaseDetailsConverter caseDetailsConverter;
+    private final CoreCaseDataService coreCaseDataService;
+    private final CamundaRestEngineClient camundaRestEngineClient;
+    private final SealedClaimFormGeneratorForSpec sealedClaimFormGeneratorForSpec;
+
+
+    @PostMapping("/generateDoc")
+    public CaseDocument uploadSealedDocument(
+        @RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation, @NotNull @RequestBody CaseData caseData) {
+        return sealedClaimFormGeneratorForSpec.generate(caseData, authorisation);
+    }
+
+
+    @PostMapping(value = "/downloadDoc",
+        produces = MediaType.APPLICATION_PDF_VALUE)
+    public @ResponseBody
+    byte[] downloadSealedDocument(
+        @RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation, @NotNull @RequestBody CaseDocument caseDocument) {
+        return sealedClaimFormGeneratorForSpec.downloadDocument(caseDocument, authorisation);
+    }
+}
