+++ src/main/java/uk/gov/hmcts/reform/civil/controllers/testingsupport/TestingSupportController.java
@@ -6,26 +6,15 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.BusinessProcess;
 import uk.gov.hmcts.reform.civil.model.CaseData;
-import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
-import uk.gov.hmcts.reform.civil.service.docmosis.sealedclaim.TestingSupportSealedClaimFormGeneratorForSpec;
-
-import javax.validation.constraints.NotNull;
 
 import static uk.gov.hmcts.reform.civil.enums.BusinessProcessStatus.STARTED;
 
@@ -39,7 +28,6 @@ public class TestingSupportController {
     private final CaseDetailsConverter caseDetailsConverter;
     private final CoreCaseDataService coreCaseDataService;
     private final CamundaRestEngineClient camundaRestEngineClient;
-    private final TestingSupportSealedClaimFormGeneratorForSpec testingSupportSealedClaimFormGeneratorForSpec;
 
     @GetMapping("/testing-support/case/{caseId}/business-process")
     public ResponseEntity<BusinessProcessInfo> getBusinessProcess(@PathVariable("caseId") Long caseId) {
@@ -60,21 +48,6 @@ public class TestingSupportController {
         }
 
         return new ResponseEntity<>(businessProcessInfo, HttpStatus.OK);
-    }
-
-
-    @PostMapping("/testing-support/case/generateDoc")
-    public CaseDocument uploadSealedDocument(
-        @RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation, @NotNull @RequestBody CaseData caseData) {
-        return testingSupportSealedClaimFormGeneratorForSpec.generate(caseData, authorisation);
-    }
-
-
-    @PostMapping(value = "/testing-support/case/downloadDoc",
-                produces = MediaType.APPLICATION_PDF_VALUE)
-    public @ResponseBody byte[] downloadSealedDocument(
-        @RequestHeader(HttpHeaders.AUTHORIZATION) String authorisation, @NotNull @RequestBody CaseDocument caseDocument) {
-        return testingSupportSealedClaimFormGeneratorForSpec.downloadDocument(caseDocument, authorisation);
     }
 
     @Data
