+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/sealedclaim/SealedClaimFormGeneratorForSpec.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.civil.service.docmosis.sealedclaim;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.ccd.document.am.model.Document;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.ClaimAmountBreakup;
 import uk.gov.hmcts.reform.civil.model.ClaimAmountBreakupDetails;
@@ -29,6 +30,7 @@ import uk.gov.hmcts.reform.civil.utils.InterestCalculator;
 import uk.gov.hmcts.reform.civil.utils.MonetaryConversions;
 
 import java.math.BigDecimal;
+import java.net.URI;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
@@ -73,6 +75,14 @@ public class SealedClaimFormGeneratorForSpec implements TemplateDataGenerator<Se
             authorisation,
             new PDF(getFileName(caseData), docmosisDocument.getBytes(), DocumentType.SEALED_CLAIM)
         );
+    }
+
+    public byte[] downloadDocument(CaseDocument caseDocument, String authorisation) {
+        String documentPath = URI.create(caseDocument.getDocumentLink().getDocumentUrl()).getPath();
+
+        byte[] file = documentManagementService.downloadDocument(authorisation, documentPath);
+        return file;
+
     }
 
     @NotNull
