+++ src/main/java/uk/gov/hmcts/reform/civil/service/documentmanagement/SecuredDocumentManagementService.java
@@ -25,12 +25,7 @@ import uk.gov.hmcts.reform.document.DocumentDownloadClientApi;
 import uk.gov.hmcts.reform.document.utils.InMemoryMultipartFile;
 import uk.gov.hmcts.reform.idam.client.models.UserInfo;
 
-import java.io.File;
-import java.io.FileOutputStream;
 import java.net.URI;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.time.ZoneId;
 import java.util.Collections;
 import java.util.Optional;
@@ -77,10 +72,6 @@ public class SecuredDocumentManagementService implements DocumentManagementServi
                 documentUploadRequest
             );
 
-            File outputFile = new File("myPdf.pdf");
-            try (FileOutputStream outputStream = new FileOutputStream(outputFile)) {
-                outputStream.write(pdf.getBytes());
-            }
             Document document = response.getDocuments().stream()
                 .findFirst()
                 .orElseThrow(() -> new DocumentUploadException(originalFileName));
@@ -116,13 +107,22 @@ public class SecuredDocumentManagementService implements DocumentManagementServi
             String userRoles = String.join(",", this.documentManagementConfiguration.getUserRoles());
             Document documentMetadata = getDocumentMetaData(authorisation, documentPath);
 
-            ResponseEntity<Resource> responseEntity = documentDownloadClientApi.downloadBinary(
+            ResponseEntity<Resource> responseEntity = caseDocumentClientApi.getDocumentBinary(
                 authorisation,
                 authTokenGenerator.generate(),
-                userRoles,
-                userInfo.getUid(),
-                URI.create(documentMetadata.links.binary.href).getPath().replaceFirst("/", "")
+                UUID.fromString(documentPath.substring(documentPath.lastIndexOf("/") + 1))
             );
+
+            if(responseEntity == null) {
+                responseEntity = documentDownloadClientApi.downloadBinary(
+                    authorisation,
+                    authTokenGenerator.generate(),
+                    userRoles,
+                    userInfo.getUid(),
+                    URI.create(documentMetadata.links.binary.href).getPath().replaceFirst("/", "")
+                );
+            }
+
 
             return Optional.ofNullable(responseEntity.getBody())
                 .map(ByteArrayResource.class::cast)
